components/edge-agent/pyproject.toml
[project]
name = "robotinc-edge-agent"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "boto3",
    "strands-agents",
    "colour",
]
